Feat : Create a page folder and move the page in it AND fix the temporary projectsPages
