+++ src/pages/App.jsx
@@ -4,10 +4,10 @@ import { useState } from 'react'
 import styled, { keyframes } from 'styled-components'
 import { SocialIcon } from 'react-social-icons'
 import { Snackbar, Alert } from '@mui/material'
-import homePagePhoto from "./assets/homePagePhoto.png"
-import ToolsSection from "./components/ToolsSection"
-import Header from './components/Header'
-import ProjectsSection from './components/ProjectsSection'
+import homePagePhoto from "../assets/homePagePhoto.png"
+import ToolsSection from "../components/ToolsSection"
+import Header from '../components/Header'
+import ProjectsSection from '../components/ProjectsSection'
 
 const name = "Hi, I'm  Charles-Eric";
 const lettersArray = name.split('');
